users-service/requirements.txt
anyio==3.6.2
attrs==22.2.0
autoflake==2.0.2
black==23.3.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.2.2
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
asyncpg==0.27.0
sqlalchemy[asyncio]==2.0.9
bcrypt==4.0.1
passlib==1.7.4
pyjwt==2.7.0
